requirements.txt
PyMuPDF
chromadb 
sentence-transformers 
openai 
streamlit
langchain 
langchain_community
pypdf

python-docx
docx2pdf
beautifulsoup4
requests
pdfplumber
python-dotenv
pypandoc
streamlit
langchain_community